refactor: Benefits.type a shared. #238
