+++ src/teacher/contexts/benefitsAPIContext/BenefitAPIContext.type.ts
@@ -1,7 +1,7 @@
 import type {
   BenefitResponseInterface,
   CreateBenefitInterface,
-} from "../../types/Benefits.type";
+} from "../../../shared/types/Benefits.type";
 import type {
   GetPaginated,
   PaginatedData,
